2.10.0

+++ sdk/src/main/java/jp/co/freee/accounting/models/PaymentRequestResponsePaymentRequest.java
@@ -380,7 +380,7 @@ public class PaymentRequestResponsePaymentRequest {
 
   public static final String SERIALIZED_NAME_TOTAL_AMOUNT = "total_amount";
   @SerializedName(SERIALIZED_NAME_TOTAL_AMOUNT)
-  private Integer totalAmount;
+  private Long totalAmount;
 
   public PaymentRequestResponsePaymentRequest() { 
   }
@@ -1208,7 +1208,7 @@ public class PaymentRequestResponsePaymentRequest {
   }
 
 
-  public PaymentRequestResponsePaymentRequest totalAmount(Integer totalAmount) {
+  public PaymentRequestResponsePaymentRequest totalAmount(Long totalAmount) {
     
     this.totalAmount = totalAmount;
     return this;
@@ -1221,12 +1221,12 @@ public class PaymentRequestResponsePaymentRequest {
   @javax.annotation.Nonnull
   @ApiModelProperty(example = "30000", required = true, value = "合計金額")
 
-  public Integer getTotalAmount() {
+  public Long getTotalAmount() {
     return totalAmount;
   }
 
 
-  public void setTotalAmount(Integer totalAmount) {
+  public void setTotalAmount(Long totalAmount) {
     this.totalAmount = totalAmount;
   }
